完善各功能点，解决绝大部分bug

+++ src/main/java/com/three/zteoa/controller/DeptController.java
@@ -9,9 +9,13 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.three.zteoa.bean.Authority;
 import com.three.zteoa.bean.Dept;
 import com.three.zteoa.bean.Emp;
 import com.three.zteoa.component.SecurityComponent;
+import com.three.zteoa.myenum.ModuleEnum;
+import com.three.zteoa.myenum.TypeEnum;
+import com.three.zteoa.service.AuthorityService;
 import com.three.zteoa.service.DeptService;
 import com.three.zteoa.vo.UpdateVo;
 
@@ -21,6 +25,10 @@ public class DeptController {
 
 	@Autowired
 	private DeptService deptService;
+	@Autowired
+	private SecurityComponent securityComponent;
+	@Autowired
+	private AuthorityService authorityService;
 
 	@RequestMapping("/queryList")
 	public List<Dept> queryList(@RequestBody(required = false) Dept dept) {
@@ -33,18 +41,39 @@ public class DeptController {
 	}
 
 	@RequestMapping("/update")
-	public boolean update(@RequestBody Dept dept) {
-		return deptService.update(dept);
+	public UpdateVo update(@RequestBody Dept dept, HttpSession session) {
+		Emp emp = (Emp) session.getAttribute("empSession");
+		UpdateVo updateVo = securityComponent.isAuthorityUpdate(emp, ModuleEnum.DEPT_MANAGER, TypeEnum.UDPATE);
+		if (updateVo.isBl()) {
+			if (deptService.update(dept)) {
+				return new UpdateVo("更新部门成功", true);
+			}
+			return new UpdateVo("更新部门失败,服务器异常", false);
+		}
+		return updateVo;
 	}
 
 	@RequestMapping("/add")
-	public boolean add(@RequestBody Dept dept) {
-		return deptService.add(dept);
+	public UpdateVo add(@RequestBody Dept dept, HttpSession session) {
+		Emp emp = (Emp) session.getAttribute("empSession");
+		UpdateVo updateVo = securityComponent.isAuthorityUpdate(emp, ModuleEnum.DEPT_MANAGER, TypeEnum.ADD);
+		if (updateVo.isBl()) {
+			if (deptService.add(dept)) {
+				return new UpdateVo("添加部门成功", true);
+			}
+			return new UpdateVo("添加部门失败,服务器异常", false);
+		}
+		return updateVo;
 	}
 
 	@RequestMapping("/delete")
-	public UpdateVo delete(Integer id) {
-		return deptService.delete(id);
+	public UpdateVo delete(Integer id, HttpSession session) {
+		Emp emp = (Emp) session.getAttribute("empSession");
+		UpdateVo updateVo = securityComponent.isAuthorityUpdate(emp, ModuleEnum.DEPT_MANAGER, TypeEnum.DELETE);
+		if (updateVo.isBl()) {
+			return deptService.delete(id);
+		}
+		return updateVo;
 	}
 
 	@RequestMapping("/queryTotal")
@@ -55,6 +84,12 @@ public class DeptController {
 	@RequestMapping("/queryByName")
 	public boolean queryByName(String name) {
 		return deptService.queryByName(name);
+	}
+
+	@RequestMapping("/getAuthoritys")
+	public List<Authority> getAuthoritys(HttpSession session) {
+		Emp emp = (Emp) session.getAttribute("empSession");
+		return authorityService.queryByEmpAndModule(emp, ModuleEnum.DEPT_MANAGER);
 	}
 
 }
